backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pyjwt==2.9.0
bcrypt==4.0.1
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
gunicorn==22.0.0
